Day 19
My inital idea was to take rule 0, iterate through its sub rules one at a time, and evaluate them against the input string. As each sub rule is satisficed, go on to the next sub rule with the remainder of the string.

Angel and Tom then pointed out that this reinvents how regex works. This is similar to yesterday's problem in that we could either reinvent regex, or use regex to solve our problem.

Using regex to solve our problem means converting rule 0 into a regex string, then match it against the input.